+++ src/lustre/vdom/reconciler.ffi.mjs
@@ -586,9 +586,7 @@ const syncedBooleanAttribute = (name) => {
 const syncedAttribute = (name) => {
   return {
     added(node, value) {
-      try {
-        node[name] = value;
-      } catch (_) {}
+      node[name] = value;
     },
   };
 };
@@ -597,8 +595,6 @@ const ATTRIBUTE_HOOKS = {
   checked: syncedBooleanAttribute("checked"),
   selected: syncedBooleanAttribute("selected"),
   value: syncedAttribute("value"),
-  width: syncedAttribute('width'),
-  height: syncedAttribute('height'),
 
   autofocus: {
     added(node) {